requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt==4.0.1]
pydantic
pydantic[email]
gitpython
python-dotenv
mysqlclient